fix: 유저 정보 페칭하는 로직을 커스텀 훅으로 분리 & RedirectPage에서 사용하지 않는 코드 제거
